pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aa-buybackprogram"
dynamic = ["version", "description"]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "Ikarus Cesaille", email = "contact@eve-linknet.com" }]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth-app-utils>=1.18.0",
    "allianceauth<5.0.0,>=4",
    "django-eveonline-sde>=0.0.1b6",
    "django-esi<10.0.0,>=8.3.1",
    "requests>=2.33.0,<3.0.0",
    "pydantic>=2.11.9,<3.0.0",
]

[project.urls]
Homepage = "https://gitlab.com/paulipa/allianceauth-buyback-program"

[tool.flit.module]
name = "buybackprogram"

[tool.isort]
profile = "black"
multi_line_output = 3
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
known_allianceauth = ["allianceauth", "app_utils"]
known_django = ["django", "django_webtest", "esi", "eve_sde"]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "faker>=40.11.1",
]
